Configurando correctamente las migraciones y la conexión a la base de datos a través de TypeORM.
