+++ .env
@@ -1,3 +1,4 @@
+NODE_ENV=development
 PORT=3001
 DB_HOST=localhost
 DB_PORT=5432
