+++ package.json
@@ -5,12 +5,12 @@
   "description": "Retro Stand Api",
   "main": "index.js",
   "scripts": {
-    "start": "ts-node-dev --env-file=.env src/index.ts",
-    "start:dist": "node --env-file=.env dist/index.js",
+    "start": "ts-node-dev src/index.ts",
+    "start:dist": "node dist/index.js",
     "build": "tsc",
     "eslint": "eslint --fix .",
-    "typeorm": "typeorm-ts-node-commonjs -d ./src/data-source/data-source.ts",
-    "m:run": "npm run typeorm migration:run"
+    "mgr:create": "typeorm migration:create src/migrations/%npm_config_name% -- -d",
+    "mgr:run": "ts-node ./node_modules/typeorm/cli.js migration:run --dataSource=./src/shared/infrastructure/persistence/migration.config.ts"
   },
   "keywords": [],
   "author": "Manuel Rivera; Jose Fuilco",
