+++ src/shared/infrastructure/persistence/migration.config.ts
@@ -1,0 +1,5 @@
+process.loadEnvFile(".env");
+
+import PostgreSQLDataSource from "./typeorm.config";
+
+export default PostgreSQLDataSource;
