+++ tsconfig.json
@@ -1,5 +1,9 @@
 {
   "compilerOptions": {
+      "baseUrl": ".",
+      "paths": {
+        "src/*": ["src/*"]
+      },
       "target": "es6",
       "module": "commonjs",
       "outDir": "./dist",
